Update style to calculator to include proper buttons instead of divs
Changed the divs that were being used as place holders for the calculor buttons
to proper button elements.

Also updates the style to change the cursor to pointer when user hovers buttons;

Still need to polish the style, but struggling with flexbox and will need a refresher
on it before updating style again. Switching to JavaScript and functionality for now