Reset script file.
Initial attempt was not working out. Deleted most of the functions/eventlisteners and started over.
Add functions to handle the numbers, operattors, and some of the calculations.
Added to the clear() function so that it would reset the number, operator, and screen global variables.

Will need to continue working on calculations and renaming the firstNum and secondNum variables to better
reflect their purpose.